Update redis example to work with Python DSL and kafka-jvm Vagrantfile
