+++ examples/redis/config.json
@@ -3,17 +3,19 @@
     "topology_specs": "topologies/",
     "virtualenv_specs": "virtualenvs/",
     "envs": {
-        "prod": {
-            "user": "",
-            "nimbus": "",
-            "workers": [],
+        "vagrant": {
+            "user": "vagrant",
+            "nimbus": "streamparse-box",
+            "workers": [
+                "streamparse-box"
+            ],
             "log": {
-                "path": "",
+                "path": "/var/log/streamparse",
                 "max_bytes": 1000000,
                 "backup_count": 10,
                 "level": "info"
             },
-            "virtualenv_root": ""
+            "virtualenv_root": "/data/virtualenvs"
         }
     }
 }
