+++ examples/redis/virtualenvs/wordcount_mem.txt
