+++ examples/redis/virtualenvs/wordcount_redis.txt
